Reorganize layout code based on layout modes
