+++ victor/src/dom/mod.rs
@@ -4,7 +4,7 @@ mod html;
 pub(crate) mod traversal;
 mod xml;
 
-use crate::style::StyleSetBuilder;
+use crate::style::{StyleSet, StyleSetBuilder};
 use html5ever::{LocalName, QualName};
 use std::borrow::Cow;
 use std::fmt;
@@ -45,7 +45,8 @@ impl Document {
         NodeId(std::num::NonZeroUsize::new(1).unwrap())
     }
 
-    pub fn parse_stylesheets(&self, style_set: &mut StyleSetBuilder) {
+    pub(crate) fn parse_stylesheets(&self) -> StyleSet {
+        let mut style_set = StyleSetBuilder::new();
         for &id in &self.style_elements {
             let element = &self[id];
             // https://html.spec.whatwg.org/multipage/semantics.html#update-a-style-block
@@ -56,6 +57,7 @@ impl Document {
             }
             style_set.add_stylesheet(&self.child_text_content(id))
         }
+        style_set.finish()
     }
 
     /// (rel_attribute, href_attribute)
