+++ victor/src/layout/flow/float.rs
@@ -1,0 +1,7 @@
+use super::*;
+
+#[derive(Debug)]
+pub(in crate::layout) struct FloatBox {
+    pub style: Arc<ComputedValues>,
+    pub contents: IndependentFormattingContext,
+}
