+++ victor/src/layout/replaced.rs
@@ -1,0 +1,4 @@
+#[derive(Debug)]
+pub(crate) enum ReplacedContent {
+    // Not implemented yet
+}
